remove_stray_courses: remove course references without organization linked
