+++ openedx/core/djangoapps/appsembler/sites/utils.py
@@ -3,13 +3,10 @@ A big module for Tahoe Sites multi-tenancy helpers.
 
 A lot of this module should be migrated into more specific modules such as `tahoe-sites`.
 """
-import tahoe_sites.api
-from django.apps import apps
 
 from datetime import timedelta
 
 import beeline
-from opaque_keys.edx.django.models import CourseKeyField, LearningContextKeyField
 
 from urllib.parse import urlparse
 
@@ -31,7 +28,6 @@ from django.utils.text import slugify
 
 from organizations import api as org_api
 from organizations import models as org_models
-from organizations.models import OrganizationCourse
 
 from organizations.models import Organization
 
@@ -45,7 +41,6 @@ from tahoe_sites.api import (
     update_admin_role_in_organization,
 )
 
-from common.djangoapps.util.organizations_helpers import get_organization_courses
 from openedx.core.lib.api.api_key_permissions import is_request_has_valid_api_key
 from openedx.core.lib.log_utils import audit_log
 from openedx.core.djangoapps.theming.helpers import get_current_request, get_current_site
@@ -53,7 +48,6 @@ from openedx.core.djangoapps.theming.models import SiteTheme
 
 from ..tahoe_tiers.legacy_amc_helpers import get_active_tiers_uuids_from_amc_postgres
 from .site_config_client_helpers import get_active_site_uuids_from_site_config_service
-from openedx.core.djangoapps.content.course_overviews.models import CourseOverview
 
 
 @beeline.traced(name="get_lms_link_from_course_key")
@@ -479,104 +473,6 @@ def bootstrap_site(site, org_data=None, username=None):
     else:
         user = {}
     return organization, site, user
-
-
-def get_models_using_course_key():
-    """
-    Get all course related model classes.
-    """
-    course_key_field_names = {
-        'course_key',
-        'course_id',
-    }
-
-    models_with_course_key = {
-        (CourseOverview, 'id'),  # The CourseKeyField with a `id` name. Hard-coding it for simplicity.
-        (OrganizationCourse, 'course_id'),  # course_id is CharField
-    }
-
-    model_classes = apps.get_models()
-    for model_class in model_classes:
-        for field_name in course_key_field_names:
-            field_object = getattr(model_class, field_name, None)
-            if field_object:
-                field_definition = getattr(field_object, 'field', None)
-                if field_definition and isinstance(field_definition, (CourseKeyField, LearningContextKeyField)):
-                    models_with_course_key.add(
-                        (model_class, field_name,)
-                    )
-
-    return models_with_course_key
-
-
-def delete_organization_courses(organization):
-    """
-    Delete all course related model instances.
-    """
-    course_keys = []
-
-    for course in get_organization_courses({'id': organization.id}):
-        course_keys.append(course['course_id'])
-
-    model_classes = get_models_using_course_key()
-
-    print('Deleting course related models:', ', '.join([
-        '{model}.{field}'.format(model=model_class.__name__, field=field_name)
-        for model_class, field_name in model_classes
-    ]))
-
-    for model_class, field_name in model_classes:
-        objects_to_delete = model_class.objects.filter(**{
-            '{field_name}__in'.format(field_name=field_name): course_keys,
-        })
-        objects_to_delete.delete()
-
-
-def remove_course_creator_role(users):
-    """
-    Remove course creator role to fix `delete_site` issue.
-
-    This will fail in when running tests from within the LMS because the CMS migrations
-    don't run during tests. Patch this function to avoid such errors.
-    TODO: RED-2853 Remove this helper when AMC is removed
-          This helper is being replaced by `update_course_creator_role_for_cms` which has unit tests.
-    """
-    from cms.djangoapps.course_creators.models import CourseCreator  # Fix LMS->CMS imports.
-    from student.roles import CourseAccessRole  # Avoid circular import.
-    CourseCreator.objects.filter(user__in=users).delete()
-    CourseAccessRole.objects.filter(user__in=users).delete()
-
-
-@beeline.traced(name="delete_site")
-def delete_site(site):
-    from third_party_auth.models import SAMLConfiguration  # local import to avoid import-time errors
-
-    print('Deleting SiteConfiguration of', site)
-    site.configuration.delete()
-
-    print('Deleting theme of', site)
-    site.themes.all().delete()
-
-    organization = tahoe_sites.api.get_organization_by_site(site)
-
-    print('Deleting users of', site)
-    users = tahoe_sites.api.get_users_of_organization(organization, without_inactive_users=False)
-    remove_course_creator_role(users)
-
-    # Prepare removing users by avoiding on_delete=models.PROTECT error
-    # SAMLConfiguration will be deleted with `site.delete()`
-    SAMLConfiguration.objects.filter(changed_by__in=users).update(changed_by=None)
-
-    users.delete()
-
-    print('Deleting courses of', site)
-    delete_organization_courses(organization)
-
-    print('Deleting organization', organization)
-    organization.delete()
-
-    print('Deleting site', site)
-    site.delete()
 
 
 @beeline.traced(name="add_course_creator_role")
